Move ActiveFilter
