+++ components/Home/SearchActiveFilters.vue
@@ -24,7 +24,7 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import ActiveFilter from '~/components/ui/ActiveFilter.vue'
+import ActiveFilter from '~/components/molecules/ActiveFilter.vue'
 import SpecIcon from '~/components/Ui/SpecIcon.vue'
 import useGuildsStore from '~/composables/useGuildsStore'
 import specializationSlug from '~/data/utils/specializationSlug'
