+++ components/atoms/icons/solid/CloseIcon.vue
